Fix ESLint violations
